modified api to return a quest by ID
