+++ test/api.test.js
@@ -30,3 +30,12 @@ test('returns lists of quests', (assert) => {
     //assert
     assert.deepEqual(quests, expected);
 });
+
+test('take in ID and return corresponding quest', (assert) => {
+    //arrange
+    const expectedQuest = questData[1];
+    //act
+    const foundQuest = api.getQuest(expectedQuest.id);
+    //assert
+    assert.deepEqual(foundQuest, expectedQuest);
+});
